requirements.txt
flask
openai
flask-cors
gunicorn
python-dotenv